requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
torch>=2.0.0
websockets>=11.0.0
python-dotenv>=1.0.0
hikvisionapi>=0.3.2  # Para conexão com DVR Hikvision (ajuste conforme seu DVR) 